+++ examples/resources/common.tsx
@@ -1,9 +1,10 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
+import { saveAs } from 'file-saver';
 
 import {
-    SerializedDiagram, WorkspaceContext,
-    ToolbarActionSave, ToolbarActionClearAll, ToolbarActionExport,
+    ElementIri, ElementModel, SerializedDiagram, WorkspaceContext,
+    ToolbarActionOpen, ToolbarActionSave, ToolbarActionClearAll, ToolbarActionExport,
 } from '../../src/index';
 
 function onPageLoad(callback: (container: HTMLDivElement) => void) {
@@ -25,6 +26,37 @@ export function ExampleToolbarMenu() {
     const {model, editor} = React.useContext(WorkspaceContext)!;
     return (
         <>
+            <ToolbarActionOpen
+                onSelect={async file => {
+                    const preloadedElements = new Map<ElementIri, ElementModel>();
+                    for (const element of model.elements) {
+                        preloadedElements.set(element.iri, element.data);
+                    }
+
+                    const json = await file.text();
+                    try {
+                        const diagramLayout = JSON.parse(json);
+                        await model.importLayout({
+                            dataProvider: model.dataProvider,
+                            diagram: diagramLayout,
+                            preloadedElements,
+                        });
+                    } catch (err) {
+                        alert('Failed to load specified file with a diagram layout.');
+                    }
+                }}>
+                Open from file
+            </ToolbarActionOpen>
+            <ToolbarActionSave mode='layout'
+                onSelect={() => {
+                    const diagramLayout = model.exportLayout();
+                    const layoutString = JSON.stringify(diagramLayout);
+                    const blob = new Blob([layoutString], {type: 'application/json'});
+                    const timestamp = new Date().toISOString().replaceAll(/[Z\s:-]/g, '');
+                    saveAs(blob, `reactodia-diagram-${timestamp}.json`);
+                }}>
+                Save to file
+            </ToolbarActionSave>
             <ToolbarActionSave mode='layout'
                 onSelect={() => {
                     const diagram = model.exportLayout();